+++ Graveyard Keeper/src/MaxButtonsRedux/GlobalUsings.cs
@@ -1,4 +1,4 @@
-﻿// Global using directives
+// Global using directives
 
 global using System;
 global using System.Collections.Generic;
@@ -6,8 +6,10 @@ global using System.Diagnostics.CodeAnalysis;
 global using System.Linq;
 global using System.Reflection;
 global using BepInEx;
+global using BepInEx.Configuration;
 global using BepInEx.Logging;
 global using HarmonyLib;
 global using Rewired;
+global using Shared;
 global using UnityEngine;
 global using Object = UnityEngine.Object;