+++ Graveyard Keeper/src/NoTimeForFishing/GlobalUsings.cs
@@ -1,8 +1,10 @@
-﻿// Global using directives
+// Global using directives
 
 global using System.Reflection;
 global using BepInEx;
+global using BepInEx.Configuration;
 global using BepInEx.Logging;
 global using Fishing;
 global using HarmonyLib;
+global using Shared;
 global using UnityEngine;