+++ Graveyard Keeper/src/MaxButtonsRedux/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## 1.3.12 | 17 April 2026
+
+- Added an update notice on the main menu that flags when this mod (or others in this collection) has a newer version on Nexus. Click an entry to open its Nexus page. Toggle off in settings if you'd rather not see it
+
 ## 1.3.11 | 12 April 2026
 
 - Renamed from "Max Button Controller Support" to "Max Buttons Redux"