+++ Graveyard Keeper/src/NoTimeForFishing/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## 3.2.9 | 17 April 2026
+
+- Added an update notice on the main menu that flags when this mod (or others in this collection) has a newer version on Nexus. Click an entry to open its Nexus page. Toggle off in settings if you'd rather not see it
+
 ## 3.2.8 | 11 April 2026
 
 - Added startup logging to help diagnose mod and game environment issues